Minor naming (capitalization) fixes
